MVC: Extract reusable `ErrorView` into a new file
